requirements.txt
Flask
Pillow
tensorflow